test(go-backend): split unit and integration coverage
