+++ compiler/transpile/transpile_test.go
@@ -1,3 +1,5 @@
+//go:build integration
+
 package transpile
 
 import (
@@ -5,113 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 	"testing"
-
-	"github.com/akonwi/ard/checker"
-	"github.com/akonwi/ard/parse"
 )
-
-func checkedModuleFromSource(t *testing.T, dir, fileName, source string) checker.Module {
-	t.Helper()
-	path := filepath.Join(dir, fileName)
-	if err := os.WriteFile(path, []byte(source), 0o644); err != nil {
-		t.Fatalf("failed to write source: %v", err)
-	}
-
-	result := parse.Parse([]byte(source), path)
-	if len(result.Errors) > 0 {
-		result.PrintErrors()
-		t.Fatalf("unexpected parse errors")
-	}
-
-	resolver, err := checker.NewModuleResolver(dir)
-	if err != nil {
-		t.Fatalf("failed to create module resolver: %v", err)
-	}
-
-	c := checker.New(fileName, result.Program, resolver)
-	c.Check()
-	if c.HasErrors() {
-		for _, diagnostic := range c.Diagnostics() {
-			t.Log(diagnostic)
-		}
-		t.Fatalf("unexpected checker errors")
-	}
-
-	return c.Module()
-}
-
-func TestEmitEntrypoint(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
-		t.Fatalf("failed to write ard.toml: %v", err)
-	}
-
-	module := checkedModuleFromSource(t, dir, "main.ard", `
-fn add(a: Int, b: Int) Int {
-  let sum = a + b
-  sum
-}
-
-let result = add(1, 2)
-`)
-
-	out, err := EmitEntrypoint(module)
-	if err != nil {
-		t.Fatalf("did not expect error: %v", err)
-	}
-
-	generated := string(out)
-	checks := []string{
-		"package main",
-		"func Add(a int, b int) int",
-		"sum := (a + b)",
-		"return sum",
-		"func main()",
-		"result := Add(1, 2)",
-		"_ = result",
-	}
-	for _, check := range checks {
-		if !strings.Contains(generated, check) {
-			t.Fatalf("expected generated source to contain %q\n%s", check, generated)
-		}
-	}
-	if strings.Contains(generated, "_ = sum") {
-		t.Fatalf("did not expect generated source to contain redundant discard for used local\n%s", generated)
-	}
-}
-
-func requireFullBuildBinaryCoverage(t *testing.T) {
-	t.Helper()
-	if fullParityEnabled() {
-		return
-	}
-	t.Skip("set ARD_FULL_PARITY=1 to run exhaustive BuildBinary coverage")
-}
-
-func TestEmitEntrypointUnusedLocalGetsDiscard(t *testing.T) {
-	dir := t.TempDir()
-	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
-		t.Fatalf("failed to write ard.toml: %v", err)
-	}
-
-	module := checkedModuleFromSource(t, dir, "main.ard", `
-fn noop() {
-  let sum = 1
-}
-
-noop()
-`)
-
-	out, err := EmitEntrypoint(module)
-	if err != nil {
-		t.Fatalf("did not expect error: %v", err)
-	}
-
-	generated := string(out)
-	if !strings.Contains(generated, "_ = sum") {
-		t.Fatalf("expected generated source to contain discard for unused local\n%s", generated)
-	}
-}
 
 func TestBuildBinaryCompilesUserModuleImport(t *testing.T) {
 	dir := t.TempDir()
@@ -151,7 +47,6 @@ let result = utils::add(1, 2)
 }
 
 func TestBuildBinaryCompilesIfReturn(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -178,7 +73,6 @@ let result = choose(2)
 }
 
 func TestBuildBinaryCompilesWhileLoop(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -205,7 +99,6 @@ let result = count(3)
 }
 
 func TestBuildBinaryCompilesForLoop(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -232,7 +125,6 @@ let result = count(3)
 }
 
 func TestBuildBinaryCompilesBreakInWhileLoop(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -287,7 +179,6 @@ let result = get_age()
 }
 
 func TestBuildBinaryCompilesImportedModuleStructLiteral(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -320,7 +211,6 @@ let result = get_age()
 }
 
 func TestBuildBinaryCompilesListLiteralAndMethods(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -345,7 +235,6 @@ let second = get_second()
 }
 
 func TestBuildBinaryCompilesListSortAndSwap(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -371,7 +260,6 @@ let value = reordered()
 }
 
 func TestBuildBinaryCompilesMutableListMethods(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -456,7 +344,6 @@ let has_none = maybe::none<Int>().is_none()
 }
 
 func TestBuildBinaryCompilesMapGetAndKeys(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -481,7 +368,6 @@ let found = lookup()
 }
 
 func TestBuildBinaryCompilesPrimitiveMethods(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -506,7 +392,6 @@ let g = true.to_str()
 }
 
 func TestBuildBinaryCompilesRangeAndForInLoops(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -564,7 +449,6 @@ let d = sum_map()
 }
 
 func TestBuildBinaryCompilesEnums(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -599,7 +483,6 @@ let b = colors::Color::green == colors::Color::red
 }
 
 func TestBuildBinaryCompilesExternStubs(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -628,7 +511,6 @@ let b = utils::meaning()
 }
 
 func TestBuildBinaryCompilesStdlibFsModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -649,7 +531,6 @@ let exists = fs::exists("./demo.txt")
 }
 
 func TestBuildBinaryCompilesStdlibIoModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -682,7 +563,6 @@ io::print(Person{name: "world"})
 }
 
 func TestBuildBinaryCompilesStdlibEncodeModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -705,7 +585,6 @@ let c = encode::json(true)
 }
 
 func TestBuildBinaryCompilesStdlibDecodeModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -748,7 +627,6 @@ let raw = decode::from_json("\{\"age\":1\}")
 }
 
 func TestBuildBinaryCompilesDecodeEndToEndFlow(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -778,7 +656,6 @@ let out = run()
 }
 
 func TestBuildBinaryCompilesGenericStdlibDecodeCombinators(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -804,7 +681,6 @@ let _ = first
 }
 
 func TestBuildBinaryCompilesStdlibArgvModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -825,7 +701,6 @@ let args = argv::load()
 }
 
 func TestBuildBinaryCompilesStdlibDatesModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -846,7 +721,6 @@ let today = dates::get_today()
 }
 
 func TestBuildBinaryCompilesStdlibDurationModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -870,7 +744,6 @@ let millis = duration::from_millis(4)
 }
 
 func TestBuildBinaryCompilesStdlibBase64Module(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -892,7 +765,6 @@ let decoded = base64::decode(encoded, true)
 }
 
 func TestBuildBinaryCompilesStdlibChronoModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -913,7 +785,6 @@ let now = chrono::now()
 }
 
 func TestBuildBinaryCompilesStdlibCryptoModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -941,7 +812,6 @@ let id = crypto::uuid()
 }
 
 func TestBuildBinaryCompilesStdlibEnvModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -962,7 +832,6 @@ let home = env::get("HOME")
 }
 
 func TestBuildBinaryCompilesStdlibFloatModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -985,7 +854,6 @@ let c = float::floor(3.9)
 }
 
 func TestBuildBinaryCompilesStdlibHexModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1007,7 +875,6 @@ let decoded = hex::decode(encoded)
 }
 
 func TestBuildBinaryCompilesStdlibIntModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1028,7 +895,6 @@ let parsed = int::from_str("42")
 }
 
 func TestBuildBinaryCompilesStdlibJsonModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1049,7 +915,6 @@ let out = json::encode(["age": 1])
 }
 
 func TestBuildBinaryCompilesStdlibMapModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1070,7 +935,6 @@ let values = map::new<Int>()
 }
 
 func TestBuildBinaryCompilesStdlibDynamicModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1118,7 +982,6 @@ let value = fiber.get()
 }
 
 func TestBuildBinaryCompilesStdlibAsyncStartModuleFunction(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1149,7 +1012,6 @@ fiber.join()
 }
 
 func TestBuildBinaryCompilesStdlibSqlModule(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1177,7 +1039,6 @@ let out = run()
 }
 
 func TestBuildBinaryCompilesStdlibSqlTransactionFlow(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1227,7 +1088,6 @@ let ok = http::Response::new(200, "ok").is_ok()
 }
 
 func TestBuildBinaryCompilesExplicitTypeArgsOnZeroArgGenericFunction(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1251,7 +1111,6 @@ let values = empty<Int>()
 }
 
 func TestBuildBinaryCompilesVariableShadowing(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1277,7 +1136,6 @@ let out = run()
 }
 
 func TestBuildBinaryCompilesStructMethods(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1316,7 +1174,6 @@ let result = run()
 }
 
 func TestBuildBinaryCompilesBoolAndIntMatch(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1351,7 +1208,6 @@ let b = bucket(2)
 }
 
 func TestBuildBinaryCompilesEnumMatch(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1382,7 +1238,6 @@ let result = label(Status::active)
 }
 
 func TestBuildBinaryCompilesMaybeCombinators(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1428,7 +1283,6 @@ let d = chained_none()
 }
 
 func TestBuildBinaryCompilesMaybeMatch(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1457,7 +1311,6 @@ let b = pick(maybe::none())
 }
 
 func TestBuildBinaryCompilesTemplateStringsAndPanic(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1484,7 +1337,6 @@ let msg = greet("Ard")
 }
 
 func TestBuildBinaryCompilesBasicResults(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1528,7 +1380,6 @@ let d = forced()
 }
 
 func TestBuildBinaryCompilesResultCombinators(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1577,7 +1428,6 @@ let d = chained_err()
 }
 
 func TestBuildBinaryCompilesResultMatches(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1785,7 +1635,6 @@ let j = nested_list(6, 3)
 }
 
 func TestBuildBinaryCompilesTryInMatchExpressions(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1869,7 +1718,6 @@ let local = multiply(3, 4)
 }
 
 func TestBuildBinaryCompilesExplicitVoidAnonymousFunctions(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1896,7 +1744,6 @@ fn main() {
 }
 
 func TestBuildBinaryCompilesContextualVoidAnonymousFunctions(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
@@ -1923,7 +1770,6 @@ fn main() {
 }
 
 func TestBuildBinaryCompilesGenericAsyncFiberLists(t *testing.T) {
-	requireFullBuildBinaryCoverage(t)
 	dir := t.TempDir()
 	if err := os.WriteFile(filepath.Join(dir, "ard.toml"), []byte("name = \"demo\"\nard = \">= 0.1.0\"\n"), 0o644); err != nil {
 		t.Fatalf("failed to write ard.toml: %v", err)
